Add files via upload
Figure S3.5: Graph displaying O2 probing conducted to confirm the level of oxygen matched what was used in the chamber in the conditions for Hyp + IR, tested before irradiation, after irradiation and at the 24 hour mark. Normoxia refers to IR alone therefore this was measured before irradiation and at the end of the 24 hours alongside Hyp alone. Hypoxic conditions were set to be 2% in the chamber, and this was confirmed with probing (1.5%) and normoxia was found to be similar to atmospheric as anticipated (19%).